fix(canon): repin attribute-name anchors and split prop sources
Repins the two remaining value-anchored expectations to the attribute
name — the check-server dependency-patch column and the Misskey LSP
shared-edit range, whose binding mode now asserts the :code attribute
name like vue-tsc — and moves the prop source-range and generated-value
helpers into expressions/prop_sources.rs to keep both files inside the
new-source length budget.
