+++ tests/performance/misskey-lsp-incremental.test.ts
@@ -151,7 +151,12 @@ test(
             });
             return waitForDiagnostics(session, componentUri, 4, true);
           });
-          assertSingleInjectedMismatch(sharedBroken.diagnostics, baseline, cleanSource, "binding");
+          assertSingleInjectedMismatch(
+            sharedBroken.diagnostics,
+            baseline,
+            cleanSource,
+            "attribute-name",
+          );
 
           const sharedRepaired = await metrics.measure("sharedRepaired", async () => {
             session.notify("textDocument/didChange", {
@@ -254,7 +259,7 @@ function assertSingleInjectedMismatch(
   diagnostics: LspDiagnostic[],
   baseline: string[],
   source: string,
-  expectedRange: "declaration" | "binding" = "declaration",
+  expectedRange: "declaration" | "attribute-name" = "declaration",
 ): void {
   const injected = diagnostics.filter(
     (diagnostic) =>
@@ -273,12 +278,11 @@ function assertSingleInjectedMismatch(
     assert.deepEqual(diagnostic.range?.start, start);
     assert.deepEqual(diagnostic.range?.end, end);
   } else {
-    // The prop-type mismatch (TS2322) anchors at the component attribute name
-    // (`code`), matching vue-tsc, rather than the authored value expression.
-    const bindingOffset = source.indexOf(`:code="String(${symbol})"`);
-    assert.notEqual(bindingOffset, -1);
-    const nameOffset = bindingOffset + ":".length;
-    const start = offsetToPosition(source, nameOffset);
+    // The child prop-type mismatch anchors at the attribute name, exactly
+    // where vue-tsc reports it.
+    const attributeOffset = source.indexOf(`:code="String(${symbol})"`);
+    assert.notEqual(attributeOffset, -1);
+    const start = offsetToPosition(source, attributeOffset + ":".length);
     const end = { line: start.line, character: start.character + "code".length };
     assert.deepEqual(diagnostic.range?.start, start);
     assert.deepEqual(diagnostic.range?.end, end);
